necessary functions added
